-Adding fixes for the bot

+++ models/hrm/hrm_act_v1.py
@@ -208,7 +208,7 @@ class HierarchicalReasoningModel_ACTV1_Inner(nn.Module):
         output = self.lm_head(z_H)[:, self.puzzle_emb_len:]
 
         # Q head
-        q_logits = self.q_head(z_H[:, 0]).to(torch.float32)
+        q_logits = self.q_head(z_H[:, 0]).to(torch.bfloat16)
         
         return new_carry, output, (q_logits[..., 0], q_logits[..., 1])
